Add files via upload
Line 445 – When registering a new user I want to distinguish between a wrong e-mail and an e-mail that was already registered with a different account. Otherwise, trying to register with the same e-mail and getting a wrong e-mail error message is kind of confusing for the user.
On line 140 is the output of the corresponding error message

Line 115 I added the template parameter to the clean_login_register shortcode. So if somebody wants to collect some extra data during registration, he is able to set up a different login form and submits the php filename as a parameter. On line 150 I changed to reflect the possibility to load a different form.

For the ability of processing the data I added a hook on line 527. I pass the newly created user as a parameter, the form data can be pulled in via the $_POST superglobal.


Line 410: The original code was destroying a session after checking the parameter. I had another plugin depending on sessions and it got messed up by that. I changed the code to just reset that one parameter instead of destroying the session